requirements.txt
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=0.19.0 